Create lily_code_iter.*, use it for closure transforming. But...
First, the good part. The old closure code is now #if 0'd out, and is
no longer being used. Instead, there's a new code iteration api that
walks over the code, which is written into an auxillary buffer in the
meantime. This makes it easier to figure out what's going on in the
transform code. The iteration is also easier and more pleasant to look
at.

The new code iteration part of the interpreter has a more flexible
structure than the old opcode table. So adding new opcodes and fixing
a couple of latent bugs (as well as disassembly) should be much, much
easier now.

Now, the bad part. Jumping back is broken, but actually was always
broken. So any closure using continue or do...while is going to break.