Fix 2+ Exception classes not working together correctly.
The problem is with dynaloading Exception before the *Error class is
completely done. Emitter assumes that the most recently entered class
is the one that will receive the class method. So what was happening
is that dynaloading was breaking that assumption. It went like this:

So what happened is that Exception would get the .new of whatever
class was dynaloading with it. That, in turn, would cause Exception to
have two different constructors: One for the original Exception, and
one for the first *Error.

What happens later is that the second Exception-derived class comes
around, but it uses the wrong constructor, so that the message and
traceback fields aren't properly set.

...and then the interpreter segfaults during raise. Good times.