Create lily_new_local_var, to replace most lily_new_scoped_var calls.
The vars of __import__ functions are implicitly globals, but they are
within their own scope instead of being in __main__. As such, they
need locals for exception capture, for loop iter, etc. This fixes a
number of strange issues/crashes that would happen when trying to
import something with code at the top.