Rewrite the outer facing part of expression handling.
Previously, there was an ast pool, and it managed itself. If you
wanted to save/restore an expression, you had to ask for a freeze
state. For a long time, that worked just fine for lambdas and even
interpolation.

For those things, it worked great. Now try rewinding that pool, and it
isn't going to be easy. You've got these freeze states that are
allocated even though they don't need to be alive at runtime, which is
another problem.

The solution is to redo it. Now, parser carries a root expression.
Think of it as a checkpoint. Expression states are created on stack
and are based off of whatever is the last checkpoint. Rewinding these
is much easier (the root becomes current, and rewind save entries).

There's also another benefit, which is that what were freeze entries
no longer hang around at runtime, and are no longer allocated. Small
gains and all.

A part of this also included changing membuf to string_pile. Many of
the functions were trimmed down, because only expressions are using
the thing (it has much more limited use than originally thought).