Lift block handling for if, try, and match.
Previously, what these blocks would do is to setup their state, and
then call down to statement. Statement would go into their peers, and
the peers would check for correctness, as well as handling their own
expression bodies. It was weird, and for if it was actually hacked
around.

Now, it's more simple and regular: The block starts up, and runs the
body it has, and then the bodies of the expressions of the peer
blocks. A huge advantage of this is that for any block created, it is
recursed into, and the block is destroyed when the recursion is
escaped out of.

I'll be using this in the next patch to do another adjustment to how
blocks work, which will make this make more sense.