Send "partial eval" to the trash with old closure code.
So the idea was that variants would produce a type instead of a result
since they may not have full type info. In the past, with a typeful
vm even, this was an odd decision. But now that the vm is typeless,
having result storages that are partials is completely harmless since
vm doesn't deal in types.

There's a rare chance that one or more functions might pull for extra
storages because of this. However, that chance is so infrequent that
if it does get hit, you won't really feel it (only a few), and it's
only going to be if you have some super huge function that builds many
sets of many partials.

What this saves is so insignificant that it's easily outweighed by the
numerous problems that have been caused by having it around. Drop it
to make the core simpler.