Drop 'enum_slot_count'. It's worse than useless.
A closer inspection of enum allocation in vm reveals a leak: Enums
were allocated for the largest variant, but then had a size of the
given variant recorded.

Also, with enums being value classes, assigning to them just creates a
new value. I initially did this because I thought that swapping out
the value of an enum class with 1 refs might be useful, but nothing in
core does that...so let's not bother, at least for now.