Lexer's entries no longer store a filename.
Lambdas, dynaload, and interpolation have made it so that lexer's list
of entries are not 1:1 with what's actually been opened. Also, a lot
of the filenames passed were actually junk so the above three could
work around that.

The most recently opened filename can be found by poking symtab's
active module's path, where that's needed.