dyna_tools.py, re-comment non-core packages (#156)
This creates dyna_tools.py, which will be the source of other tools
that all use the same docblock comment base. All non-core packages
have been commented. The core, which is big, will be up next.

dynascan.py, as well as the tool directory it was in are now gone, as
they were not very useful to begin with.