SYM_CLASS_TAINTED added, and always give Tainted that id.
This is a matter of consistency: Tainted was the only built-in class
that needed an entry in the cid table. It seems reasonable to give
Tainted a particular id since there's a very small cost in doing so.

mod_lily updated to no longer use a cid table for Tainted.