BIG! Refcounts start at 0 now, moves do a ref increase.
This isn't many lines, but it's a BIG change to the internals of how
refcounts go. Previously, a newly-made value started at one ref, and a
move would put it somewhere. That worked for a long time. But then I
started to write the new API and I noticed a problem.

Start with String.split. It creates a List, and it needs to put
Strings inside. Ideally, Lists are created pre-sized, and with
elements ready to receive a value. If I wanted a function to transfer
the value into a List, I would write lily_list_set_string.

But...the string starts at 1 ref, so it doesn't need a ref bump.

But lily_list_set_value, would need to bump a ref.

I could write lily_list_set_string_ref, for cases where I want to set
a specific kind of value, but also want a ref.

That's way too easy to get wrong. So the new approach is that assign
AND move now both increase the refcount. Newly-made raw values start
at 0 refs to offset that (except functions, which never get a true
move or assign).