Classes aren't related to everything before them (#232).
In the test provided, Alpha knew at emit-time that it had no relation
to Beta. At runtime, Alpha incorrectly thought that Beta was a parent,
and thus re-used Beta's structure.

This later caused @s to be overridden by @a. Now, the String that was
to be passed to upper would instead be an instance! The mere fact that
@a (a tagged value, for other reasons) was inside of Beta (a
non-tagged value) could also crash the vm during a gc pass, as @a
would be incorrectly reclaimed.