Big! Create a new and proper API for the backend.
This patch is a massive change to how the backend is organized, and
how foreign functions interact with the interpreter. The result of
this is a mostly (but not entirely) complete foundation for Lily going
forward.

Layout:

* lily_move.* has been created, and holds all move operations. These
  will be phased out shortly.

* Create lily_value_structs.* for parts of the interpreter that need
  to know about the internals of a value.

* Create lily_api_value.* which contain value operations and which
  present raw Lily values as opaque pointers. This has resulted in the
  addition of helper functions to grab parts of said values. Expect
  the helper functions to be expanded upon in the near future.

* lily_api_value_flags.h has been created to hold the flags describing
  a value.

The intent of this patch is to provide functions such that foreign
functions do not need to know the intrinsics of a value. That effort
is still a work in progress, but will eventually allow postgres and
apache to use only api files.

Content:

* Foreign functions now use lily_arg_* functions to receive values
  from the interpreter. No more argc, no more code *.

* To do the above, functions internally now have their values inside
  of registers instead of pulling them from the caller's stack. There
  is some work to be done cleaning up how that works, but this is
  about establishing a foundation usable by others.

* Calling back into the interpreter is now done by pushing values onto
  the interpreter's stack. Following a call, the interpreter rewinds
  the stack to what it was before the call. However, the values of a
  call all dump to a single register held by the caller. Values can
  now accumulate on the stack via pushing more than what's needed,
  then later popping it away.

* lily_list_set_* functions now exist for setting values of different
  types onto a particular element of a list. These same functions
  exist for pushing values, setting a variant's content, an instance's
  content, and so forth. This means that, overall, there is less
  reliance on moves, and less of a need to interact with the guts of
  how the gc works.

* In addition, there are now lily_return_* functions that handle
  returning a value. Now, a foreign function only needs to take the vm
  pointer, and is able to use the new api functions to do as it needs
  to.

The API for the backend is by no means complete. But this does a large
amount of what needs to be done, and allows others to start working
with it.

The intent of this patch is, overall, to make extensions to the
intepreter agnostic to the layout of values that are inside of it. The
hope is that, one day, there might be a LilyJIT that can seamlessly
load extensions written for stock Lily.