Replace lily_new_list_val with the new _n suffixed one.
This also drops lily_new_string_take in favor of
lily_new_raw_string_take, with the new version pushing back a raw
string instead of one held in a value. That seems to be the trend now.

As for lists, this removes the old 'constructor' that would give a
hollow list, and then expect the caller to fill it in. It's a hair
slower because the values are set and then have to be moved. But it's
well worth it in encapsulation. This also reduces list constructors to
just one.