Big: Variants are now proper classes.
This was done to make it easier to discover what variant that an enum
is without having to dig the enum itself. That will allow me to
hopefully drop the variant members field of an enum some time in the
future.

This also makes it easier to build enums. Before, one had to know the
position of the enum, as well as what instance that the enum fell
under. Now, each variant gets a unique id so a library would only need
to know just the id of the variant. That means less mucking around
with cids. A small bonus, but still a bonus.

Some parts of the interpreter had to change:

o_match_dispatch now passes in the class id + 1, so that the instance
id can be subtracted by it. This allows the same jumping as before,
but with a different way of getting the value.

Classes had their move_flags moved (ha) so that id could be brought up
and aligned with variant classes. Variant classes now get a proper id
instead of the offset from the start that they used to have.

Most importantly, the padding that previously existed for enums and
variants in their struct is still there. It's possible to shave a
whole uint64_t, but doing so destroys alignment with lists. I've
decided, for the moment, to leave the padding in, but I may change
that later. If I do, the API should not be damaged from it.