api: Include msgbuf, but encapsulate it.
msgbuf is now fully encapsulated: The struct now lives in the .c file,
so that the .h only mentions what operations are available. I've
changed the prefix to just mb from msgbuf so that it's not as wordy to
use.

The full encapsulation comes with the vm having a new lily_api_vm.h so
that I could make lily_vm_msgbuf to get (and flush!) the vm's msgbuf
as part of the api.

I've also taken grow out of msgbuf's api. Exporting that was a bad
idea, and callers should instead do what File.read_line does instead
(caller has a fixed-size buffer, sends sized chunks at once).

I plan to trim msgbuf's api, but not by too much. Some functions only
make sense for vm (like adding integers or bytestrings). But right now
I want to focus on doing the last bit that postgres so that I can push
postgres out of the core.