Emitter: Don't write unreachable branch exit jumps.
The real reason I'm doing this is that a branch that always exits has
exit jumps that target the next code position. If all branches exit,
then there should not be any code after the branches.

This causes a problem during closure transform, because now there's a
jump to a position for which there is no bytecode position.

So let's not do this altogether.