Relative jumps have finally landed.
This is a slight but powerful change to how jumps are processed.
Before, the jump position that was written was always an absolute
position. For the longest time, that didn't matter because the vm had
a code_pos at runtime that it would add all the time to code.

But then the vm dropped that awful idea in favor of doing direct
adjustments to code. It turns out that not doing repeated adds for
every code access is actually NOT FAST! Who'd have thought, eh?

Absolute jumps threw a wrench into that. It required calculating the
distance code had moved away from the starting position, or getting
the starting code position and calculating a move against that.

Relative jumps mean that the moved code can have a value directly
added to it. It unfortunately doesn't make much of a difference with
speed though.

I also fixed what was almost certainly a bug: An attempt to do a
transform was being done on o_variant_decompose. I don't typically
like to mix fixes, but it was preventing this from working and I'm
REALLY sure that's wrong. REALLY sure.