Fix not being able to close over a for loop var (fixes #254).
This was caused by not writing the target down as an output. Fixing
that yielded the wrong number. So I did some adjustment, and it turns
out that the for loop op itself didn't have the right ordering either.
So I fixed that too. If I hadn't, the number would have been one lower
than the start of the loop.

I've added a test to make sure this doesn't come back.