Re-enable closure optarg tests (fixes #183).
When I did relative jumps, I fixed o_variant_decompose having a count
wrongly transformed, and I believe that may have been the source of
problems with that insofar as closures go.

What this does is make it so that optional arguments and closures work
together. I had to go and fix optarg writing, because that assumed
optargs is always first (o_new_instance, at least, can come before).

The problem in the closure transform part is that optargs weren't
being ensured in the closure. It was actually employing almost reverse
logic (keeping optarg params out). That's definitely wrong: Any and
all closed-over params need a force o_set_upvalue to make sure they
are present in the closure.