Fix potential crash when importing a file with a toplevel for loop.
function depth is one for __main__'s block as well as __import__'s
block. That's because __import__'s vars are globals, but the storages
are not. So if __import__ had a for loop, there would be a potential
crash because syncing assigns would be written to a register that was
local instead of global.