Create lily_generic_pool to manage generic state.
This isn't a big change, but it's an important one. Before this
commit, Lily had one true generic class. Slight of hand was used to
place different generic -types- onto the one generic -class-. This,
combined with parser trying to hand-manage generic state caused
quite a few problems.

The way generics was implemented was a barrier to many features such
as nameable generics, System F, etc. It was also getting in the way
of rewind, and it was ugly. I do not miss the old strategy.

If you declare generics in one scope, generics in the next scope now
always start where the old ones left off. Previously, that didn't
work with enums, and I patched around it before instead of fixing it
(which is why tests are changed with this).