Create lily_api_embed.h.
I've decided to create a base lily_state as a resulting typedef for
the parse state. This means that the interpreter's internals can
change later on (perhaps exporting a vm as the interpreter's front
if the internals are compiler-driven?)

I've created specific functions for template use. My plan is for
templates to be executed and be done, but for regular runs to be
setup for additional actions (calling specific globals or somesuch).
I haven't added that functionality in yet though.