api: Export only lily_state now, hopefully.
This unifies the api for both extending and embedding the interpreter
so that it only need know about a lily_state. This is directly
inspired by Lua, and a want for uniformity across the api.

Bear in mind that most operations within lily_api_value.h are not
suitable for use outside of extending the interpreter. But a majority
of them are made for that purpose.