api: By default, have parser free options during shutdown.
Most of the time, options can be thrown away when the parser is done,
so this makes that the default. Some experimentation and user feedback
will be done in the future to make sure that it's clear to users that
they shouldn't free options at exit, or maybe if another strategy is
better.