Lay down the foundation of rewind (fixes #174). However...
I've tested this to pass all the failing tests when run twice. It's
rough around the edges, and I may have missed some parts. First, I'd
like to talk about the bad part:

It's very, very difficult for an interpreter to do a full, fresh
rewind. No matter what, there are some parts which must stick around.
One example, is if the interpreter retains a class that used data that
was recently loaded. It's then harder to rewind away functions,
because rewind may accidentally rewind away data that's necessary.

So, with that in mind, the current rewind strategy is extremely
conservative, in favor of holding onto extra data versus potentially
releasing something that may cause a crash. This means that, for the
time being, the interpreter slowly leaks memory if the vm fails during
exec and the last pass included some classes or extra functions.

Literals are not being rewound either, so it's possible for an
interpreter to accumulate them as well. I intend to eventually roll
out a better, more complete rewind that scans through the vm to
determine what's absolutely necessary. With code iter, I believe it's
possible to be perfect with rewinding.

But for now, rewinding isn't perfect. But it is good. This is about
establishing a foundation that others can use. This opens the door to
having a working repl, and I've been itching to have that for some
time now.