Show `use` and packages the door (fixes #272).
Distinguishing between packages and modules does not add value, when
they are so similar on the outside. This commit scraps hopefully
everything about packages: It's now back to just modules.

`import` now does what `use` did in the past. That is, it will look at
a relative path, then at a "packages" directory. This time, however,
the package directory is rooted from the first module that is loaded.

I thought, briefly, about having a relative packages directory that
would allow overriding of the rooted package directory. I decided
against it, as it's easier to add it later if there's demand, than to
pull it if it turns out to be worthless (something something spacebar
heating).