Allow 'self' as a return type (fixes #138).
I'm not going to force people into using it, because I don't think
this feature is a hard sell. I also wrote in a rule to allow 'self'
to be implicitly returned like with functions that don't return a
value.

I have NOT added support for 'self' to dynaload yet, but I will in the
near future.