Add sys.getenv (fixes #216).
It's a wrapper over C's getenv, to allow access to environment
information. Note that it currently assumes that the result is a valid
`String` value. In the future, the wrapper might check for utf-8
correctness (and give `None` if that's the case). But...we'll get to
that when we get there.