Disable implicit Integer <-> Double conversions.
I noticed this recently when I went to start adding Byte to the
language. My reasoning boils down to "I don't think this is useful and
I don't like it." If someone has a use or a good reason for it, then
I'll add it back in post-1.0.

It's easier to add features in later than to take them out.