BIG! lily_value now stores the class id in the lower 16 bits.
Previously, lily_value stored bitflags for each of the predefined
classes. User-defined classes stored their real class id within an
'instance_id' field.

Now, the flags field of lily_value stores markers in the upper 16 bits
(out of 32). The lower 16 stores the exact class id for all values.
This has in not needing to store the class id inside of instance and
enum values.

This breaks instance and enum-related api. Everything else should be
okay. This change was done for the following reasons:

* Empty variants needed 'default' backing values, thus discouraging
  the use of large enums. I'd like to keep the overhead small.

* On the C side, getting default backing values is difficult. It
  required looking up the class identity from vm's class table, then
  doing about 7 pointer accesses to get the default value. This made
  returning `None` a tad irritating. Packages wishing to work with
  their own empty variants would be difficult too.

* Previously, move flags were used to translate a class to the flags
  it had at vm time. Move flags were coming up to the limit of a 16
  bit int. I was forced to either expand and pad out lily_class, or
  take a different approach.

The new approach has the drawback that you can't glean the class id of
an instance or variant from the raw value.

I've added and adjusted api functions in accordance with the shift in
internals. I've also increased the id of all classes by one. That was
necessitated by how flags being 0 indicates the absence of a value.
If any class has id 0 and it's primitive, there's no way to tell it
apart from an unset value.