Make dyna_tools.py easier to extend, and slightly adjust format.
The only adjustment to the format is to make `m:` into `m\0`. The
interpreter doesn't care what the value is (it's for alignment), but
this makes it more consistent.

The main changes overall is to make the class generating parts of
dyna_tools pull more weight, and to flatten dynaload/loader
generation. These two changes make it easier to extend dyna_tools.