api: Replace dynaload bootstrap with native dynaload (fixes #282).
This allows dynaload to introduce native classes to the interpreter.
These classes can be inherited from, because they really are the same
kind of classes that you'd build natively.

I've added an api function: lily_ctor_setup. You'll call it if you
have a foreign function acting as the constructor for a native class
(what this needed). It handles the "is this part of an inherited class
or is it being directly called" question for you. You only need to set
the properties where they belong.

A side-effect of this is that the documentation will be more verbose
as these dynaloaded classes now have dynaloaded constructors (seemed
reasonable to have that be in two parts).