Lambdas use parentheses now (fixes #290)
The bulk of this patch is rewriting lambdas to use `(|<vars>| ...)`
instead of `{|<vars>| ... }`. The reasons I've done this are:

The driving force behind this patch is consistency. It's easier to
close expressions when they only contain parentheses (`))` instead of
`})`). It's also consistent, I feel, because curly braces now focus
on blocks, whereas expressions use only parentheses.

I've made a slight adjustment to lexer: Lambda and interpolation
loading now yield a corresponding end token instead of a generic eof.
This was done because lexer can have mismatched curly braces now, and
in such cases parser will pump back the proper unexpected token.