syntax: Dynamic has pointer equality now.
A long time ago, when Dynamic was called Any (and further back, when
it was Object), it was a more transparent container. Values assigned
to Any would implicitly get wrapped into Any, so this "made sense".

I used to call that feature "defaulting to Any". I didn't like it.
Dynamic is more explicit, and I want to make it a necessary evil. So
give it sucky pointer equality, because this is the #1 way to have
comparison throw RuntimeError. Mind you, it's still possible for that
to happen, but you need some very, very deep containers.