api: Either -> Result, lily_file_for_*, value api shuffling.
Many people have rightfully (heh) mentioned in the past that Either
ought to be called Result instead. That Left and Right are more
baggage and not names that really convey a purpose. I was waiting
for a good time to do this, and I think this is it. So Either, Left,
and Right are now Result, Failure, and Success (respectively).

More central to api, lily_file_ensure_* functions are gone in place
of lily_file_for_*, which produce the raw file on success.

The last two changes are:

* lily_push_super -> lily_instance_super, and put by
  lily_new_instance to encourage proper use.

* lily_is_valid_utf8 moved to be by the string handling functions.