cmake: -O2 by default for non-Windows, -DQUICK_BUILD added.
I suspect that most people interested in trying Lily for the first
time will implicitly want the faster executable. -O2 makes a pretty
big difference in benchmarks, so it should be the default.

I've added a QUICK_BUILD flag on the non-Windows side, which can be
set to on to skip using -O2. I'll be using this, because I'd rather
have the faster builds. I'd add it to the Windows side too, but I
don't know the relevant flag, and I don't use Windows.