api: lily_get_msgbuf now lily_get_clean/dirty_msgbuf.
A small problem I've had on occasion is the instinctual need to clean
a msgbuf right after getting it. I know that lily_get_msgbuf does it
for me, and yet I forget and call to clear it twice. It's a very small
problem, but I'd like to solve it before 1.0.

This removes the _noflush suffixed version, replacing both with the
two above. I've done an additional step, having the dirty_msgbuf
return to a pointer, to prevent accidentally calling it and getting a
dirty buffer.