Lexer: Property words (@<name>) have to be reloaded too.
This is a hack, because lexer shouldn't actually be working like it is
as far as entry push/pop. But this fixes cases where a property word
after a dynaload would fail with a strange error message. In the
example added, it previously complained about property A not being in
class Test when trying to use the property.