Allow 'match' to work against classes (fixes #304).
With this patch, it is now possible to use 'match' against monomorphic
user-defined classes. Prior to this, it was not possible to extract
a derived class from a var typed as being the base. Whoops.

For now, pattern matching will assign over to a new var that will have
the proper class.

```
class One {}
class Two < One {}

var v: One = Two

match v: {
    case Two(t):
        ...
    else:
        ...
}
```