tm: When building types, only keep the essential flags.
Prior to this, some sources were sending in types which had a base
like TYPE_IS_UNRESOLVED. That would then carry over to the new type
(which may BE resolved). This shows up very obviously when unable to
capture, say, an `Option[Integer]` and being met with an error about
how it isn't resolved.