Expr: Fix binary merge using possibly the wrong parent.
In the case that `tree` has climbed upward, `active->parent` is not
equal to `parent`, so make sure to use `parent` here. This doesn't
break any tests, and I couldn't get it to break with the current code.

However, in testing the new `++` concat feature, I found that a
construct like `x ++ y ++ z + a ++ b ++ c` (subtle single plus mark in
the center) can cause an infinite loop in climbing trees. The reason,
it turns out, is that the parent was improperly set, which this patch
fixes.