Fix a nasty bug where hash values could become "invisible".
Hash values internally carry a pointer to their raw key, which is
linked in lifetime to their boxed key. When a hash value was updated,
the old raw key wasn't updated. Since that raw key is used for
comparisons, the hash would have a key that couldn't be updated.

This is moderately difficult to trigger: You need a `Hash` that has
type `String`, the `String` needs to not be interned, and the `String`
backing value needs to be deref'd/gc swept away.