BIG: Now using `bindgen.lily` for generating bindings.
Lily's documentation and binding generation is a mess. This is a step
toward cleaning that mess up. What this does is to have the bindings
that Lily uses internally now be generated by a `bindgen.lily` script
kept in a different repo. That tool relies on a parsing engine that
can understand Lily definitions, which is a step toward other neat
tools.

The old tool, `dyna_tools.py` will be kept until the new tool is able
to generate documentation. The new tool's input is pretty close to the
old tool's expectations, so migrating over won't be that difficult.

This also adjusts Lily's internal loading mechanism. Previously, it
always looked for `lily_dynaload_table`. Now, it looks for
`lily_<packagename>_table` as well. That means that multiple tables
can live together in a single executable without clobbering the same
name.