Lift render func + data away from lexer into parser.
"in template" information is now passed through parser as arguments,
so that it doesn't need to be saved in a structure somewhere. I think
that's the right move.

More importantly, this removes lexer from needing to know about
configuration/render func/data. Because now parser has a template
reading loop to handle that.