api: Config now includes an import hook (fixes #233).
The interpreter now exports a set of lily_open_* functions that take
some path (and possibly extra data), and attempt to load it. The
interpreter will check if a duplicate name has already been loaded, as
well as remembering what paths are tried. This allows it to print the
same "no file named ..." error message at end for any paths that the
loader tries.

The lily_open_* set of functions also include `lily_load_library`
which abstracts over dlopen/LoadLibraryA, getting the dynaload table,
etc.

The import hook is called after the interpreter tries to load modules
that are builtin/registered to it. There's still no way of configuring
off builtin modules.

The sandbox ifdef gate around import has been removed entirely,
allowing the sandbox to now make use of builtin imports. The sandbox
has an import hook that's a no-op, because it doesn't have a true
underlying filesystem.