Remove class-level closures (fixes #311).
Previously, class constructors allowed lambdas to close over `self`,
which allowed them to modify class members. It was also permissible to
close over parameters inside of a constructor. In either case, a
closure would be created that would be stored in the class behind the
scenes.

It is no longer permissible to close over the parameters of a class
constructor. Users who want that can instead assign the parameter to
a private class member instead. Closing over `self` is now forbidden
as well.

Now, if a class method is a closure, it can only close over content
that is inside of it. As a result, there is no longer a need to create
closures at the class level. The opcode o_load_class_closure has been
done away with.

Some of this is about simplifying closures. But this also partially
addresses the problem of a potentially incomplete `self` (such as in
the case of an exception in a constructor) being stored somewhere
global.