Function values for closures now store their locals.
Previously, the emitter would write o_load_closure at the top of any
non-backing closure function. The instruction would copy the upvalues,
then drop the locals from upvalues. Dropping locals from a closure is
necessary to make sure that closure cells are fresh.

The idea was right, but the implementation was not. Calling a closure
multiple times would lead to multiple drops of closure inputs. I
strongly believe that this may have been a large contributor to many
of the closure problems that Lily has faced.

Now, functions that are closures (and not the backing closure) have
the upvalues cleared out in o_create_function. In doing this, upvalues
are only cleared once. Furthermore, upvalues are now stored and
unstored from function values themselves instead of on call frames.