Rewrite a large part of the value api (fixes #310).
The idea for this rewrite came about from taking a peek at the value
api after working on the configuration structure.

The following changes have been made:

* Users can no longer directly create values. Instead, value creation
  is done by using `lily_push_*`. When pushing a container, the
  container is returned so it can be filled. As before, the language
  expects that the container is filled when returning.

* Functions have been added to set a `Hash`/container value from the
  stack.

* Given the above, `lily_box_*` functions are now gone.

* Creating superclasses is now more straightforward: `lily_push_super`
  to put it on the stack, then `lily_return_super` at the end.

* Reduce the number of return functions. Most functions will now end
  with a single extra value on the stack, and can thus use the new
  `lily_return_top` to send it back.

What's important about this is that it prevents the user from creating
a value without using the state object. That opens the door to future
possibilities like a custom allocator...if I ever go that route.

I apologize for breaking the api so severely after a 1.0 release.
There may be small adjustments to the api in the future, but none as
severe as this.

For the time being, the documentation will be left stale. I'm planning
a few other breaking changes in the near future, and I'd prefer to
update the documentation all at once.