Lexer: Add backslash at end of string feature from Rust.
If a backslash is found at the end of a string, no newline will be
written in. Additionally, space characters at the start of the string
will not be counted. This allows creating triple-quote strings that
are indented without that indentation showing up in output.

This works for both bytestring and string literals. No testing was
added, because a new testing framework will be coming soon.