Create a new lily-driven test suite (fixes #293).
This creates a new test suite that is driven by Lily code that uses
subinterpreters. The heart of the suite is a test object that other
test files add test cases to. The new test suite is composed of old
tests but merged into fewer files through use of subinterpreters.

For the moment, this involves the creation of a different executable,
since I can't quite convince CMake to build a library in the test
directory for the plain Lily executable to load.

The test suite is now run by using `pre-commit-tests` from the same
directory in which it resides. This will run the actual test suite.
The suite is now driven by Lily code.

The downside of this test suite is that a crashing test will stop the
suite. Also, if the interpreter is severely broken, the suite will
simply not function.

On the other hand, the test suite running under a single process makes
it easier to test through valgrind and to eventually get coverage
metrics.