Calls now put their result at the end of their inputs.
Calls started doing this a long time ago because there was no api for
foreign functions. So they had to grab registers directly. For the
return to be in a consistent spot, it had to be first. That's no
longer the case, so calls now put their result at the end. This makes
calls consistent with other opcodes, allowing some future
simplification.