Emitter: Remove call states from call handling.
Call states were a nice step forward when they first were added. They
provided a way for calls to not need to modify trees during eval. But
as time has gone on, call states have not been so great. They're big,
their structure is a mess, and much of the information they contain
can be gained from poking the source ast the right way.

Now, calls work by having a main call body that runs arguments as well
as handling the error part as well.

A nice side-effect of this is that call states will no longer be
stored on the stack. The call value buffer has also been disposed of.