lily_u16_insert => lily_u16_set_at.
The above is a macro that performs what's effectively a subscript
assignment (`insert(x, y, z) = x[y] = z`). There's also a function
called inJECT that inserts a value and slides all the other values to
the right.

The results of using the last one are obvious (a certain crash).
However, trying to figure out the cause can be very difficult since
the names are so similar.

Yes, that did happen just now. That's why I'm fixing this.