Expression state now stores checkpoints in an array.
Previously, expression states were created on the stack. This meant
that only one expression could be saved at a time. Also, it meant
that an entire expression state was kept on stack when it was kept.

This creates an array of checkpoints so that checkpoint status is
kept within the expression state and not on the stack. A very
conservative default size of 2 is used because they aren't created
that often.