ts: Don't solve for scoop types in resolve.
Resolve functions are used to map from generic types to what they've
been solved as. The only time scoop types need to truly be replaced is
when the resulting argument is being calculated.

For now, I've simply removed this functionality since no functions
remain that return a scoop type. Once I have a library to allow for
testing the api, I'll throw in a test for that.