Stack line numbers are now obtained through code[-1].
Frames no longer need to carry a 'line number' field on them. If the
vm is about to execute a call that may fail, then the current frame's
code is fixed beforehand. This makes it so that code[-1] is valid for
all native function frames.

As a result, the vm's 'pending line' field has been completely erased
from the interpreter.